changing customer panel
