+++ App.config
@@ -4,6 +4,6 @@
 		<clear />
     <add name="Server"
 		 providerName="System.Data.ProviderName"
-		 connectionString="Data Source=elbrus.liara.cloud,32970;Initial Catalog=myDB;User Id=sa;Password=r6at7itfz2dU8apysewdh6uv;Encrypt=False" />
+		 connectionString="Data Source=(localdb)\MSSQLLocalDB;Initial Catalog=master;Integrated Security=True;Connect Timeout=30;Encrypt=False;Trust Server Certificate=False;Application Intent=ReadWrite;Multi Subnet Failover=False" />
 	</connectionStrings>
 </configuration>
