+++ Models/Complaint.cs
@@ -17,12 +17,12 @@ namespace Restaurant_Manager.Models
         [Key, Required]
         public int Id { get; set; }
         [Required]
-        public string Details { get; set; }
+        public string Detail { get; set; }
         [Required]
         public string Title { get; set; }
         public string? Answer { get; set; }
         [Required]
-        public virtual User Users { get; set; }
+        public virtual User User { get; set; }
         [Required]
         public virtual Restaurant? Restaurant { get; set; }
         [Required]
