+++ Restaurant_Manager.csproj
@@ -16,4 +16,8 @@
     <PackageReference Include="Microsoft.EntityFrameworkCore.SqlServer" Version="8.0.6" />
   </ItemGroup>
 
+  <ItemGroup>
+    <Folder Include="CustomerPannle\" />
+  </ItemGroup>
+
 </Project>
